update imports, rename and client name
